feat: remove node dependencies from compiler
